move exchangeCodeForTokens to corsair/core
